spring core funcionando, base de datos

+++ src/frgp/utn/edu/ar/dao/IdaoMedico.java
@@ -10,7 +10,7 @@ import frgp.utn.edu.ar.entidad.Medico;
 		public boolean add(Medico medico);
 		public Medico readOne(String medico);
 		public List<Medico> readAll();
-		public boolean exist(long id);
+		public boolean exist(int id);
 		public boolean update(Medico medico);
 		public boolean delete(Medico medico);
 	}
