+++ src/frgp/utn/edu/ar/negocio/IMedicoNegocio.java
@@ -9,7 +9,7 @@ public interface IMedicoNegocio {
 	public boolean add(Medico medico);
 	public Medico readOne(String medico);
 	public List<Medico> readAll();
-	public boolean exist(long id);
+	public boolean exist(int id);
 	public boolean update(Medico medico);
 	public boolean delete(Medico medico);
